# Conflicts:
#   client/src/main/java/nl/tudelft/oopp/group39/controllers/BuildingSceneController.java
#   client/src/main/java/nl/tudelft/oopp/group39/controllers/MainSceneController.java
#   client/src/main/java/nl/tudelft/oopp/group39/models/Booking.java
#   client/src/main/java/nl/tudelft/oopp/group39/views/UsersDisplay.java

+++ server/src/main/java/nl/tudelft/oopp/group39/building/exceptions/BuildingNotFoundException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.building.exceptions;
 
 public class BuildingNotFoundException extends RuntimeException {
-    public BuildingNotFoundException(int id) {
+    public BuildingNotFoundException(Long id) {
         super("Room with id " + id + " wasn't found.");
     }
 }
