+++ server/src/main/java/nl/tudelft/oopp/group39/room/services/RoomService.java
@@ -4,12 +4,10 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-
 import nl.tudelft.oopp.group39.facility.entities.Facility;
 import nl.tudelft.oopp.group39.facility.services.FacilityService;
 import nl.tudelft.oopp.group39.room.dao.RoomDao;
 import nl.tudelft.oopp.group39.room.entities.Room;
-import nl.tudelft.oopp.group39.room.exceptions.RoomExistsException;
 import nl.tudelft.oopp.group39.room.exceptions.RoomNotFoundException;
 import nl.tudelft.oopp.group39.room.repositories.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,8 +23,8 @@ public class RoomService {
     @Autowired
     private FacilityService facilityService;
 
-    public Room readRoom(long id) throws RoomNotFoundException {
-        return roomRepository.findById(id).orElseThrow(() -> new RoomNotFoundException((int) id));
+    public Room readRoom(Long id) throws RoomNotFoundException {
+        return roomRepository.findById(id).orElseThrow(() -> new RoomNotFoundException(id));
     }
 
     public List<Room> listRooms() {
@@ -43,8 +41,8 @@ public class RoomService {
     /**
      * Doc. TODO Sven
      */
-    public Room updateRoom(Room newRoom, int id) throws RoomNotFoundException {
-        return roomRepository.findById((long) id)
+    public Room updateRoom(Room newRoom, Long id) throws RoomNotFoundException {
+        return roomRepository.findById(id)
             .map(room -> {
                 newRoom.setId(id);
                 room = newRoom;
@@ -68,7 +66,7 @@ public class RoomService {
     /**
      * Doc. TODO Sven
      */
-    public Room deleteRoom(int id) throws RoomNotFoundException {
+    public Room deleteRoom(Long id) throws RoomNotFoundException {
         try {
             Room rf = readRoom(id);
             roomRepository.delete(readRoom(id));
